notifications_service/requirements.txt
mailjet-rest
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-Migrate==4.0.4
psycopg2-binary==2.9.3
Werkzeug==2.2.2
psycopg2-binary==2.9.3
pulsar-client==3.3.0
fastavro
cryptography==42.0.0
flask-limiter
